Move endpoints to their modules.
