+++ Backend/Communication/Send.hs
@@ -10,29 +10,32 @@ import           Network.Wai.Handler.Warp
 import           Servant
 
 import           Course.File
-import qualified Course.List                         as CL
+import           Course.List
 import           Files.State
-import           Peek.Peek
-import           Settings.Exception.GeneralException
+import qualified Peek.Peek                as P
 import           Settings.Monad.Global
-
-type Excepted a = Either SomeException a
-type ExDisplayed a = Either String a
-
-type Sync = [(CL.Course, DownloadSummary)]
-type Courses = [CL.Course]
 
 type API = "sync" :> Get '[JSON] (ExDisplayed Sync)
       :<|> "courses" :> Get '[JSON] (ExDisplayed Courses)
-      :<|> "peek" :> Capture "path" FilePath :> Get '[JSON] PeekResponse
+      :<|> "peek" :> Capture "path" FilePath :> Get '[JSON] P.PeekResponse
 
 server :: Server API
-server = liftIO sync :<|> liftIO courses :<|> (liftIO . peekAt)
+server = liftIO sync :<|> liftIO courses :<|> (liftIO . P.peekAt)
+
+startApp :: IO ()
+startApp = run 8080 app
+
+app :: Application
+app = serve api server
+
+api :: Proxy API
+api = Proxy
+
 
 courses :: IO (ExDisplayed Courses)
 courses = do
     putStrLn "endpoint: \"courses\""
-    leftToString <$> runGlobal CL.thisTermCourse
+    leftToString <$> runGlobal thisTermCourse
 
 sync :: IO (ExDisplayed Sync)
 sync = do
@@ -50,41 +53,7 @@ loopListening channel asyncR = do
     this = loopListening channel asyncR
     reshape = return . joinEither
 
-runGlobal :: Global a -> IO (Excepted a)
-runGlobal g = do
-    (envR, envS) <- setup
-    (\(a,_,_) -> a) <$> runRWST (runExceptT g) envR envS
-
-runGlobalAsync :: Global a -> IO (EnvS, Async (Excepted a))
-runGlobalAsync g = do
-    (envR, envS) <- setup
-    asynced <- async $ (\(a,_,_) -> a) <$> runRWST (runExceptT g) envR envS
-    return (envS, asynced)
-
-syncG :: Global Sync
-syncG = do
-    this <- CL.thisTermCourse
-    defPath <- lift $ asks getDefaultPath
-    states <- mapM (downloadFromCourse defPath) this
-    return $ zip this states
-
-setup :: IO (EnvR, EnvS)
-setup = do
-    maybeEnvR <- tryGetEnvR
-    envR <- maybe (error "Config read fault!") return maybeEnvR
-    envS <- getEnvS
-    return (envR, envS)
-
 joinEither :: Either e (Either e a) -> Either e a
 joinEither (Left ex)           = Left ex
 joinEither (Right (Left ex))   = Left ex
 joinEither (Right (Right val)) = Right val
-
-startApp :: IO ()
-startApp = run 8080 app
-
-app :: Application
-app = serve api server
-
-api :: Proxy API
-api = Proxy
